refactor(log): move log interfaces to domain/log/interfaces
Consolidate log-related interfaces into a dedicated directory for better organization and maintainability. This change affects multiple files that import these interfaces.
