+++ src/domain/log/interfaces/log.interfaces.ts
