+++ src/infrastructure/datasources/log.datasource.ts
@@ -4,8 +4,8 @@ import { CreateLogDto } from "@/domain/log/dtos/create-log.dto";
 import { DatabaseClient } from "../config/database.client";
 import { BadRequestError } from "@/domain/errors/bad-request-error";
 import { ValidationError } from "@/domain/errors/validation-error";
-import type { LogLevel } from "@/domain/interfaces/log.interfaces";
-import { LOG_LEVELS } from "@/domain/interfaces/log.interfaces";
+import type { LogLevel } from "@/domain/log/interfaces/log.interfaces";
+import { LOG_LEVELS } from "@/domain/log/interfaces/log.interfaces";
 import { TRawJson } from "@/domain/interfaces/general.interfaces";
 import { $Enums } from "@prisma/client";
